Move tests to unexport module
