+++ tests/export.rs
@@ -1,5 +1,3 @@
-use super::*;
-
 test! {
   name:     success,
   justfile: r#"
@@ -176,98 +174,4 @@ test! {
   args: ("A=zzz"),
   stdout: "undefined\n",
   stderr: "echo $B\n",
-}
-
-#[test]
-fn unexport_environment_variable_linewise() {
-  #[cfg(not(target_os = "macos"))]
-  let stderr_msg = "echo $JUST_TEST_VARIABLE\nbash: line 1: JUST_TEST_VARIABLE: unbound variable\nerror: Recipe `recipe` failed on line 4 with exit code 127\n";
-  #[cfg(target_os="macos")]
-  let stderr_msg = "echo $JUST_TEST_VARIABLE\nbash: JUST_TEST_VARIABLE: unbound variable\nerror: Recipe `recipe` failed on line 4 with exit code 127\n";
-
-  Test::new()
-    .justfile(
-      "
-     unexport JUST_TEST_VARIABLE
-
-     recipe:
-         echo $JUST_TEST_VARIABLE
-      ",
-    )
-    .env("JUST_TEST_VARIABLE", "foo")
-    .stderr(stderr_msg)
-    .status(127)
-    .run();
-}
-
-#[test]
-fn unexport_environment_variable_shebang() {
-  Test::new()
-    .justfile(
-      "
-     unexport JUST_TEST_VARIABLE
-
-     recipe:
-         #!/usr/bin/env bash
-         echo \"variable: $JUST_TEST_VARIABLE\"
-      ",
-    )
-    .env("JUST_TEST_VARIABLE", "foo")
-    .stdout("variable: \n")
-    .status(0)
-    .run();
-}
-
-#[test]
-fn duplicate_unexport_fails() {
-  Test::new()
-    .justfile(
-      "
-     unexport JUST_TEST_VARIABLE
-
-     recipe:
-         echo \"variable: $JUST_TEST_VARIABLE\"
-
-     unexport JUST_TEST_VARIABLE
-      ",
-    )
-    .env("JUST_TEST_VARIABLE", "foo")
-    .stderr("error: Variable `JUST_TEST_VARIABLE` is unexported multiple times\n ——▶ justfile:6:10\n  │\n6 │ unexport JUST_TEST_VARIABLE\n  │          ^^^^^^^^^^^^^^^^^^\n")
-    .status(1)
-    .run();
-}
-
-#[test]
-fn export_unexport_conflict() {
-  Test::new()
-    .justfile(
-      "
-     unexport JUST_TEST_VARIABLE
-
-     recipe:
-         echo \"variable: $JUST_TEST_VARIABLE\"
-
-     export JUST_TEST_VARIABLE := 'foo'
-      ",
-    )
-    .stderr("error: Variable JUST_TEST_VARIABLE is both exported and unexported\n ——▶ justfile:6:8\n  │\n6 │ export JUST_TEST_VARIABLE := 'foo'\n  │        ^^^^^^^^^^^^^^^^^^\n")
-    .status(1)
-    .run();
-}
-
-#[test]
-fn unexport_doesnt_override_local_recipe_export() {
-  Test::new()
-    .justfile(
-      "
-     unexport JUST_TEST_VARIABLE
-
-     recipe $JUST_TEST_VARIABLE:
-         @echo \"variable: $JUST_TEST_VARIABLE\"
-      ",
-    )
-    .args(["recipe", "value"])
-    .stdout("variable: value\n")
-    .status(0)
-    .run();
 }
